requirements.txt
sqlalchemy
flask
flask_sqlalchemy
flask_login
pillow